fix(imx93): add null terminator to mmap region array
The mmap_add() API iterates through the memory region array until it
encounters a zero entry. Without a proper null terminator {0}, the
behavior is undefined.

When TF-A runs on OCRAM, the memory is zeroed by default, masking
this bug. However, when loaded by QNX SPL (which may place TF-A in
non-zeroed memory), the missing terminator causes boot failure.

Add the required {0} terminator to bl_regions[] array.

Signed-off-by: Maximus <maximus.sun@nxp.com>
Change-Id: If7f1313bcb04d97ac760e27984508ba4e6c6c752
Signed-off-by: Jacky Bai <ping.bai@nxp.com>